+++ src/test/java/org/dependencytrack/tasks/MetricsUpdateTaskIT.java
@@ -1,3 +1,21 @@
+/*
+ * This file is part of Dependency-Track.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ * Copyright (c) Steve Springett. All Rights Reserved.
+ */
 package org.dependencytrack.tasks;
 
 import alpine.persistence.JdoProperties;
@@ -35,7 +53,7 @@ import java.util.UUID;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-public class NewMetricsUpdateTaskIT {
+public class MetricsUpdateTaskIT {
 
     @Ignore // There is no image for arm64
     public static class MssqlIT {
@@ -69,7 +87,7 @@ public class NewMetricsUpdateTaskIT {
 
         @Test
         public void test() {
-            new NewMetricsUpdateTask().inform(new MetricsUpdateEvent(MetricsUpdateEvent.Type.PORTFOLIO));
+            new MetricsUpdateTask().inform(new MetricsUpdateEvent(MetricsUpdateEvent.Type.PORTFOLIO));
 
             try (final var qm = new QueryManager()) {
                 assertResult(qm.getMostRecentPortfolioMetrics());
@@ -110,7 +128,7 @@ public class NewMetricsUpdateTaskIT {
 
         @Test
         public void test() {
-            new NewMetricsUpdateTask().inform(new MetricsUpdateEvent(MetricsUpdateEvent.Type.PORTFOLIO));
+            new MetricsUpdateTask().inform(new MetricsUpdateEvent(MetricsUpdateEvent.Type.PORTFOLIO));
 
             try (final var qm = new QueryManager()) {
                 assertResult(qm.getMostRecentPortfolioMetrics());
